requirements.txt
psycopg2-binary==2.9.3
pycairo==1.20.1
pydbus==0.6.0
PyGObject==3.42.0
PyYAML==6.0
SQLAlchemy==1.4.29
